Connection details are added
In addition to drop down list wtih server names, text field with connection details is added. The text field shows host:port while the tooltip is also present user:password. It is possible to change connection details right in the toolbar. In such situation default AuthenticationMechansim is used. If user/passwords are missing, they are also taken from default settings. This will allow to quickly connect to a server without addint it to the list.