Persist position and size of Server List frame
Server List window appears in the same place where it was closed previous time in the Studio window. In a new Studio window, Server List will be opened with the same size and the same location relative to the top left corner of the Studio window. If Server List can't be fit into the screen with the described algorithm, it will be opened with default position ans size. Relative location and size of Server List is persited into config file.