Syntax highlighting in output result and other improvements
Non list output will have q syntax highlighting. It is useful if output is a function.
Added configured limit of characters in output. There are two settings - one for table cells and another one for the whole output.
Table cells are aligned to left.
Fix soring of strings and general lists.
Byte vectors are not format as list.